refactor(coordination): split types and helpers out of __init__
Move SignalListener + the CoordinationBackend alias to coordination/types.py and the
close_pubsub helper to coordination/utils.py, mirroring the superset/distributed_lock
layout (types.py + utils.py) and the existing coordination/exceptions.py. __init__
re-exports the public names, so `from superset.coordination import ...` is unchanged.
